+++ lib/core/util/tool_func/func/http.dart
@@ -1,78 +1,4 @@
-import 'dart:convert';
-
-import 'package:dart_openai/dart_openai.dart';
-import 'package:dio/dio.dart';
-import 'package:fl_lib/fl_lib.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:gpt_box/data/model/chat/history.dart';
-import 'package:gpt_box/data/res/l10n.dart';
-import 'package:html/parser.dart' as html_parser;
-
-part 'tool.p.dart';
-
-abstract final class OpenAIFuncCalls {
-  static const _tools = [
-    _HttpReq(),
-    //_RunJS(),
-  ];
-
-  static final tools = _tools.map((e) => e.into).toList();
-
-  static Future<_Ret?> handle(
-    _CallResp resp,
-    ToolConfirm askConfirm,
-    OnToolLog onToolLog,
-  ) async {
-    final tool = tools.firstWhere((t) => t.type == resp.type);
-    switch (tool.type) {
-      case 'function':
-        final fn = tool.function;
-        final args = await _parseMap(resp.function.arguments);
-        final func = _tools.firstWhereOrNull((e) => e.name == fn.name);
-        if (func == null) throw 'Unknown function ${fn.name}';
-        if (!await askConfirm(func, func.help(resp, args))) return null;
-        return await func.run(resp, args, onToolLog);
-      default:
-        throw 'Unknown tool type ${tool.type}';
-    }
-  }
-}
-
-abstract final class ToolFunc {
-  final String name;
-  final String? description;
-  final _Map parametersSchema;
-
-  const ToolFunc({
-    required this.name,
-    this.description,
-    required this.parametersSchema,
-  });
-
-  /// For users to understand what this function does.
-  /// Used in permission request dialog.
-  String help(_CallResp call, _Map args) {
-    return '''
-
-${json.encode(call)}
-
-${json.encode(args)}
-''';
-  }
-
-  Future<_Ret> run(_CallResp call, _Map args, OnToolLog log) async =>
-      throw 'Not implemented';
-
-  OpenAIToolModel get into => OpenAIToolModel(
-        type: 'function',
-        function: OpenAIFunctionModel(
-          name: name,
-          description: description,
-          parametersSchema: parametersSchema,
-        ),
-      );
-}
+part of '../tool.dart';
 
 final class _HttpReq extends ToolFunc {
   const _HttpReq()
@@ -206,23 +132,3 @@ If blob, encode it into base64 String.''',
     return ChatContent.text(respBody).asList;
   }
 }
-
-// final class _RunJS extends _ToolFunc {
-//   const _RunJS()
-//       : super(
-//           name: 'runJS',
-//           description: 'Run JavaScript code',
-//           parametersSchema: const {'code': 'The JavaScript code to run'},
-//         );
-
-//   @override
-//   Future<_Msg> run(_CallResp call, _Map args) async {
-//     final code = args['code'];
-//     return _Msg(
-//       role: _Role.tool,
-//       content: [ChatContent.text('running $code').toOpenAI],
-//       name: name,
-//       toolCalls: [call],
-//     );
-//   }
-// }
