+++ lib/core/util/tool_func/type.dart
