+++ lib/view/page/home/home.dart
@@ -11,7 +11,7 @@ import 'package:flutter/material.dart';
 import 'package:gpt_box/data/model/chat/history.share.dart';
 import 'package:gpt_box/core/route/page.dart';
 import 'package:gpt_box/core/util/chat_title.dart';
-import 'package:gpt_box/core/util/tool.dart';
+import 'package:gpt_box/core/util/tool_func/tool.dart';
 import 'package:gpt_box/core/util/image.dart';
 import 'package:gpt_box/core/util/sync/base.dart';
 import 'package:gpt_box/data/model/chat/config.dart';
